# docs/08_backend/zipsa_openapi.yaml
openapi: 3.0.3
info:
  title: ZIPSA API
  description: AI 기반 고양이 품종 매칭 및 케어 상담 서비스 백엔드 API
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: 로컬 개발 서버
  - url: https://api.zipsa.app/api/v1
    description: 프로덕션 서버
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js 발급 JWT 토큰
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "리소스를 찾을 수 없습니다."

    # ── 유저 프로필 ───────────────────────────────────
    UserPreferences:
      type: object
      description: 온보딩 선호 데이터 (MongoDB user_profiles.preferences)
      properties:
        housing:
          type: string
          enum: [apartment, house, studio]
        activity:
          type: string
          enum: [low, medium, high]
        experience:
          type: string
          enum: [beginner, intermediate, expert]
        work_style:
          type: string
          nullable: true
        allergy:
          type: boolean
        has_children:
          type: boolean
        has_dog:
          type: boolean
        traits:
          type: array
          items:
            type: string
        companion:
          type: array
          items:
            type: string
    UserProfileResponse:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        nickname:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        gender:
          type: string
          enum: [M, F, 미설정]
          nullable: true
        contact:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
          description: "OCI Object Storage 공개 URL. 없으면 OAuth 프로필 이미지 사용."
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        onboarding_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfileCreateRequest:
      type: object
      required: [preferences]
      properties:
        nickname:
          type: string
        age:
          type: integer
          nullable: true
        gender:
          type: string
          enum: [M, F, 미설정]
          nullable: true
        contact:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserProfileUpdateRequest:
      type: object
      properties:
        nickname:
          type: string
        age:
          type: integer
          nullable: true
        gender:
          type: string
          enum: [M, F, 미설정]
          nullable: true
        contact:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        preferences:
          $ref: '#/components/schemas/UserPreferences'
...
paths:
...
  /users/me/profile:
    get:
      tags: [User Profile]
      summary: 내 프로필 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: 프로필 없음 (온보딩 미완료)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [User Profile]
      summary: 프로필 생성 (온보딩 완료)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileCreateRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '409':
          description: 이미 프로필 존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [User Profile]
      summary: 프로필 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
...